fix: segmentation crash for kconf cc
When you do kconf cc for a kubeconfig file, where is context reffering
to non-existing cluster, the app will crash. That's because I always
expected the existing cluster - not broken configuration.

Anyway even broken configuration should somehow work. I've fixed it by
checking if cluster exist. If not, then the empty URL is used in list.